// client/models/UpdateUserDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateUserDto = {
    /**
     * ID of the user to update
     */
    userId: string;
    /**
     * Full name of the user
     */
    name?: string;
    /**
     * Jordanian mobile number (must start with 07)
     */
    mobileNumber?: string;
};