// client/models/ResetPasswordDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ResetPasswordDto = {
    /**
     * Jordanian mobile number of the user whose password to reset
     */
    mobileNumber: string;
    /**
     * New password for the account (minimum 6 characters)
     */
    newPassword: string;
};